Public release of MSCL 41.0.8

+++ MSCL/source/mscl/MicroStrain/Wireless/DatalogDownloader.h
@@ -145,7 +145,7 @@ namespace mscl
         //    - <Error_NotSupported>: Logged data is not supported by the Node.
         //    - <Error_NodeCommunication>: Failed to read info from the Node.
         //    - <Error_Connection>: A connection error has occurred with the parent BaseStation.
-        DatalogDownloader(const WirelessNode& node, uint16 startAddress, uint32 size);
+        DatalogDownloader(const WirelessNode& node, uint32 startAddress, uint32 size);
 
         ~DatalogDownloader();
